[AzL3Tdnf] Refactor: Addressing PR feedback #1
