+++ src/core/tests/Test_AzL3TdnfPackageManager.py
@@ -16,7 +16,6 @@
 import json
 import os
 import unittest
-import sys
 # Conditional import for StringIO
 try:
     from StringIO import StringIO  # Python 2
@@ -39,17 +38,11 @@ class TestAzL3TdnfPackageManager(unittest.TestCase):
         self.runtime.stop()
 
     # region Mocks
-    def mock_do_processes_require_restart_raise_exception(self):
-        raise Exception
-
     def mock_linux_distribution_to_return_azure_linux(self):
         return ['Microsoft Azure Linux', '3.0', '']
 
     def mock_linux_distribution_to_return_azure_linux_2(self):
         return ['Common Base Linux Mariner', '2.0', '']
-
-    def mock_write_with_retry_raise_exception(self, file_path_or_handle, data, mode='a+'):
-        raise Exception
 
     def mock_run_command_output_return_tdnf_3(self, cmd, no_output=False, chk_err=True):
         """ Mock for run_command_output to return tdnf 3 """
@@ -87,84 +80,6 @@ class TestAzL3TdnfPackageManager(unittest.TestCase):
         return '2.9.0'
     # endregion
 
-    # region Utility Functions
-    def __setup_config_and_invoke_revert_auto_os_to_system_default(self, package_manager, create_current_auto_os_config=True, create_backup_for_system_default_config=True, current_auto_os_update_config_value='', apply_updates_value="",
-                                                                   download_updates_value="", enable_on_reboot_value=False, installation_state_value=False, set_installation_state=True):
-        """ Sets up current auto OS update config, backup for system default config (if requested) and invoke revert to system default """
-        # setup current auto OS update config
-        if create_current_auto_os_config:
-            self.__setup_current_auto_os_update_config(package_manager, current_auto_os_update_config_value)
-
-        # setup backup for system default auto OS update config
-        if create_backup_for_system_default_config:
-            self.__setup_backup_for_system_default_OS_update_config(package_manager, apply_updates_value=apply_updates_value, download_updates_value=download_updates_value, enable_on_reboot_value=enable_on_reboot_value,
-                                                                    installation_state_value=installation_state_value, set_installation_state=set_installation_state)
-
-        package_manager.revert_auto_os_update_to_system_default()
-
-    def __setup_current_auto_os_update_config(self, package_manager, config_value='', config_file_name="automatic.conf"):
-        # setup current auto OS update config
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, config_file_name)
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, config_value)
-
-    def __setup_backup_for_system_default_OS_update_config(self, package_manager, apply_updates_value="", download_updates_value="", enable_on_reboot_value=False, installation_state_value=False, set_installation_state=True):
-        # setup backup for system default auto OS update config
-        package_manager.image_default_patch_configuration_backup_path = os.path.join(self.runtime.execution_config.config_folder, Constants.IMAGE_DEFAULT_PATCH_CONFIGURATION_BACKUP_PATH)
-        backup_image_default_patch_configuration_json = {
-            "dnf-automatic": {
-                "apply_updates": apply_updates_value,
-                "download_updates": download_updates_value,
-                "enable_on_reboot": enable_on_reboot_value
-            }
-        }
-        if set_installation_state:
-            backup_image_default_patch_configuration_json["dnf-automatic"]["installation_state"] = installation_state_value
-        self.runtime.write_to_file(package_manager.image_default_patch_configuration_backup_path, '{0}'.format(json.dumps(backup_image_default_patch_configuration_json)))
-
-    @staticmethod
-    def __capture_std_io():
-        # arrange capture std IO
-        captured_output = StringIO()
-        original_stdout = sys.stdout
-        sys.stdout = captured_output
-        return captured_output, original_stdout
-
-    def __assert_std_io(self, captured_output, expected_output=''):
-        output = captured_output.getvalue()
-        self.assertTrue(expected_output in output)
-
-    def __assert_reverted_automatic_patch_configuration_settings(self, package_manager, config_exists=True, config_value_expected=''):
-        if config_exists:
-            reverted_dnf_automatic_patch_configuration_settings = self.runtime.env_layer.file_system.read_with_retry(package_manager.dnf_automatic_configuration_file_path)
-            self.assertTrue(reverted_dnf_automatic_patch_configuration_settings is not None)
-            self.assertTrue(config_value_expected in reverted_dnf_automatic_patch_configuration_settings)
-        else:
-            self.assertFalse(os.path.exists(package_manager.dnf_automatic_configuration_file_path))
-    # endregion
-
-    def test_do_processes_require_restart(self):
-        """Unit test for tdnf package manager"""
-        # Restart required
-        self.runtime.set_legacy_test_type('HappyPath')
-        package_manager = self.container.get('package_manager')
-        self.assertTrue(package_manager)
-        self.assertTrue(package_manager.is_reboot_pending())
-
-        # Restart not required
-        self.runtime.set_legacy_test_type('SadPath')
-        package_manager = self.container.get('package_manager')
-        self.assertTrue(package_manager is not None)
-        self.assertFalse(package_manager.is_reboot_pending())
-
-        # Fake exception
-        self.runtime.set_legacy_test_type('SadPath')
-        package_manager = self.container.get('package_manager')
-        self.assertTrue(package_manager is not None)
-        backup_do_processes_require_restart = package_manager.do_processes_require_restart
-        package_manager.do_processes_require_restart = self.mock_do_processes_require_restart_raise_exception
-        self.assertTrue(package_manager.is_reboot_pending())  # returns true because the safe default if a failure occurs is 'true'
-        package_manager.do_processes_require_restart = backup_do_processes_require_restart
-
     def test_all_classification_selected_for_auto_patching_request(self):
         """Unit test for tdnf package manager for auto patching request where all classifications are selected since Azure Linux does not have classifications"""
         backup_envlayer_platform_linux_distribution = LegacyEnvLayerExtensions.LegacyPlatform.linux_distribution
@@ -187,324 +102,6 @@ class TestAzL3TdnfPackageManager(unittest.TestCase):
         self.assertTrue(execution_config.included_classifications_list == [Constants.PackageClassification.CRITICAL, Constants.PackageClassification.SECURITY, Constants.PackageClassification.OTHER])
 
         LegacyEnvLayerExtensions.LegacyPlatform.linux_distribution = backup_envlayer_platform_linux_distribution
-
-    def test_disable_auto_os_updates_with_uninstalled_services(self):
-        # no services are installed on the machine. expected o/p: function will complete successfully. Backup file will be created with default values, no auto OS update configuration settings will be updated as there are none
-        self.runtime.set_legacy_test_type('SadPath')
-        package_manager = self.container.get('package_manager')
-        package_manager.disable_auto_os_update()
-        self.assertTrue(package_manager.image_default_patch_configuration_backup_exists())
-        image_default_patch_configuration_backup = json.loads(self.runtime.env_layer.file_system.read_with_retry(package_manager.image_default_patch_configuration_backup_path))
-        self.assertTrue(image_default_patch_configuration_backup is not None)
-
-        # validating backup for dnf-automatic
-        self.assertTrue(package_manager.dnf_auto_os_update_service in image_default_patch_configuration_backup)
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_download_updates_identifier_text], "")
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_apply_updates_identifier_text], "")
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_enable_on_reboot_identifier_text], False)
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_installation_state_identifier_text], False)
-
-    def test_disable_auto_os_updates_with_installed_services(self):
-        # all services are installed and contain valid configurations. expected o/p All services will be disabled and backup file should reflect default settings for all
-        self.runtime.set_legacy_test_type('HappyPath')
-        package_manager = self.container.get('package_manager')
-
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = yes\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-
-        package_manager.disable_auto_os_update()
-        self.assertTrue(package_manager.image_default_patch_configuration_backup_exists())
-        image_default_patch_configuration_backup = json.loads(self.runtime.env_layer.file_system.read_with_retry(package_manager.image_default_patch_configuration_backup_path))
-        self.assertTrue(image_default_patch_configuration_backup is not None)
-
-        # validating backup for dnf-automatic
-        self.assertTrue(package_manager.dnf_auto_os_update_service in image_default_patch_configuration_backup)
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_download_updates_identifier_text], "yes")
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_apply_updates_identifier_text], "yes")
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_enable_on_reboot_identifier_text], False)
-        self.assertEqual(image_default_patch_configuration_backup[package_manager.dnf_auto_os_update_service][package_manager.dnf_automatic_installation_state_identifier_text], True)
-
-    def test_disable_auto_os_update_failure(self):
-        # disable with non existing log file
-        package_manager = self.container.get('package_manager')
-
-        self.assertRaises(Exception, package_manager.disable_auto_os_update)
-        self.assertTrue(package_manager.image_default_patch_configuration_backup_exists())
-
-    def test_update_image_default_patch_mode(self):
-        package_manager = self.container.get('package_manager')
-        package_manager.os_patch_configuration_settings_file_path = package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-
-        # disable apply_updates when enabled by default
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = yes\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-
-        package_manager.update_os_patch_configuration_sub_setting(package_manager.dnf_automatic_apply_updates_identifier_text, "no", package_manager.dnf_automatic_config_pattern_match_text)
-        dnf_automatic_os_patch_configuration_settings_file_path_read = self.runtime.env_layer.file_system.read_with_retry(package_manager.os_patch_configuration_settings_file_path)
-        self.assertTrue(dnf_automatic_os_patch_configuration_settings_file_path_read is not None)
-        self.assertTrue('apply_updates = no' in dnf_automatic_os_patch_configuration_settings_file_path_read)
-        self.assertTrue('download_updates = yes' in dnf_automatic_os_patch_configuration_settings_file_path_read)
-
-        # disable download_updates when enabled by default
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = yes\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.os_patch_configuration_settings_file_path, dnf_automatic_os_patch_configuration_settings)
-        package_manager.update_os_patch_configuration_sub_setting(package_manager.dnf_automatic_download_updates_identifier_text, "no", package_manager.dnf_automatic_config_pattern_match_text)
-        dnf_automatic_os_patch_configuration_settings_file_path_read = self.runtime.env_layer.file_system.read_with_retry(package_manager.os_patch_configuration_settings_file_path)
-        self.assertTrue(dnf_automatic_os_patch_configuration_settings_file_path_read is not None)
-        self.assertTrue('apply_updates = yes' in dnf_automatic_os_patch_configuration_settings_file_path_read)
-        self.assertTrue('download_updates = no' in dnf_automatic_os_patch_configuration_settings_file_path_read)
-
-        # disable apply_updates when default patch mode settings file is empty
-        dnf_automatic_os_patch_configuration_settings = ''
-        self.runtime.write_to_file(package_manager.os_patch_configuration_settings_file_path, dnf_automatic_os_patch_configuration_settings)
-        package_manager.update_os_patch_configuration_sub_setting(package_manager.dnf_automatic_apply_updates_identifier_text, "no", package_manager.dnf_automatic_config_pattern_match_text)
-        dnf_automatic_os_patch_configuration_settings_file_path_read = self.runtime.env_layer.file_system.read_with_retry(package_manager.os_patch_configuration_settings_file_path)
-        self.assertTrue(dnf_automatic_os_patch_configuration_settings_file_path_read is not None)
-        self.assertTrue('download_updates' not in dnf_automatic_os_patch_configuration_settings_file_path_read)
-        self.assertTrue('apply_updates = no' in dnf_automatic_os_patch_configuration_settings_file_path_read)
-
-    def test_update_image_default_patch_mode_raises_exception(self):
-        package_manager = self.container.get('package_manager')
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = yes\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-        self.runtime.env_layer.file_system.write_with_retry = self.mock_write_with_retry_raise_exception
-        self.assertRaises(Exception, package_manager.update_os_patch_configuration_sub_setting)
-
-    def test_get_current_auto_os_patch_state_with_uninstalled_services(self):
-        # no services are installed on the machine. expected o/p: function will complete successfully, backup file is not created and function returns current_auto_os_patch_state as disabled
-        self.runtime.set_legacy_test_type('SadPath')
-        package_manager = self.container.get('package_manager')
-        package_manager.get_current_auto_os_patch_state = self.runtime.backup_get_current_auto_os_patch_state
-        current_auto_os_patch_state = package_manager.get_current_auto_os_patch_state()
-
-        self.assertFalse(package_manager.image_default_patch_configuration_backup_exists())
-        self.assertEqual(current_auto_os_patch_state, Constants.AutomaticOSPatchStates.DISABLED)
-
-    def test_get_current_auto_os_patch_state_with_installed_services_and_state_disabled(self):
-        # dnf-automatic is installed on the machine. expected o/p: function will complete successfully, backup file is NOT created and function returns current_auto_os_patch_state as disabled
-        self.runtime.set_legacy_test_type('HappyPath')
-        package_manager = self.container.get('package_manager')
-        package_manager.get_current_auto_os_patch_state = self.runtime.backup_get_current_auto_os_patch_state
-
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = no\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-
-        current_auto_os_patch_state = package_manager.get_current_auto_os_patch_state()
-
-        self.assertFalse(package_manager.image_default_patch_configuration_backup_exists())
-        self.assertEqual(current_auto_os_patch_state, Constants.AutomaticOSPatchStates.DISABLED)
-
-    def test_get_current_auto_os_patch_state_with_installed_services_and_state_enabled(self):
-        # dnf-automatic is installed on the machine. expected o/p: function will complete successfully, backup file is NOT created and function returns current_auto_os_patch_state as enabled
-
-        # with enable on reboot set to false
-        self.runtime.set_legacy_test_type('HappyPath')
-        package_manager = self.container.get('package_manager')
-        package_manager.get_current_auto_os_patch_state = self.runtime.backup_get_current_auto_os_patch_state
-
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = yes\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-
-        current_auto_os_patch_state = package_manager.get_current_auto_os_patch_state()
-
-        self.assertFalse(package_manager.image_default_patch_configuration_backup_exists())
-        self.assertEqual(current_auto_os_patch_state, Constants.AutomaticOSPatchStates.ENABLED)
-
-        # with enable on reboot set to true
-        self.runtime.set_legacy_test_type('AnotherSadPath')
-        package_manager = self.container.get('package_manager')
-        package_manager.get_current_auto_os_patch_state = self.runtime.backup_get_current_auto_os_patch_state
-
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = no\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-
-        current_auto_os_patch_state = package_manager.get_current_auto_os_patch_state()
-
-        self.assertFalse(package_manager.image_default_patch_configuration_backup_exists())
-        self.assertEqual(current_auto_os_patch_state, Constants.AutomaticOSPatchStates.ENABLED)
-
-    def test_get_current_auto_os_patch_state_with_installed_services_and_state_unknown(self):
-        # dnf-automatic is installed on the machine. expected o/p: function will complete successfully, backup file is NOT created and function returns current_auto_os_patch_state as unknown
-
-        self.runtime.set_legacy_test_type('HappyPath')
-        package_manager = self.container.get('package_manager')
-        package_manager.get_current_auto_os_patch_state = self.runtime.backup_get_current_auto_os_patch_state
-
-        package_manager.dnf_automatic_configuration_file_path = os.path.join(self.runtime.execution_config.config_folder, "automatic.conf")
-        dnf_automatic_os_patch_configuration_settings = 'apply_updates = abc\ndownload_updates = yes\n'
-        self.runtime.write_to_file(package_manager.dnf_automatic_configuration_file_path, dnf_automatic_os_patch_configuration_settings)
-
-        current_auto_os_patch_state = package_manager.get_current_auto_os_patch_state()
-
-        self.assertFalse(package_manager.image_default_patch_configuration_backup_exists())
-        self.assertEqual(current_auto_os_patch_state, Constants.AutomaticOSPatchStates.UNKNOWN)
-
-    def test_revert_auto_os_update_to_system_default(self):
-        revert_success_testcase = {
-            "legacy_type": 'HappyPath',
-            "stdio": {
-                "capture_output": False,
-                "expected_output": None
-            },
-            "config": {
-                "current_auto_update_config": {
-                    "create_current_auto_os_config": True,
-                    "current_auto_os_update_config_value": 'apply_updates = no\ndownload_updates = no\n'
-                },
-                "backup_system_default_config": {
-                    "create_backup_for_system_default_config": True,
-                    "apply_updates_value": "yes",
-                    "download_updates_value": "yes",
-                    "enable_on_reboot_value": True,
-                    "installation_state_value": True,
-                    "set_installation_state": True
-                }
-            },
-            "assertions": {
-                "config_value_expected": 'apply_updates = yes\ndownload_updates = yes\n',
-                "config_exists": True
-            }
-        }
-
-        revert_success_with_dnf_not_installed_testcase = {
-            "legacy_type": 'SadPath',
-            "stdio": {
-                "capture_output": False,
-                "expected_output": None
-            },
-            "config": {
-                "current_auto_update_config": {
-                    "create_current_auto_os_config": False,
-                    "current_auto_os_update_config_value": ''
-                },
-                "backup_system_default_config": {
-                    "create_backup_for_system_default_config": True,
-                    "apply_updates_value": "",
-                    "download_updates_value": "",
-                    "enable_on_reboot_value": False,
-                    "installation_state_value": False,
-                    "set_installation_state": True
-                }
-            },
-            "assertions": {
-                "config_value_expected": "",
-                "config_exists": False
-            }
-        }
-
-        revert_success_with_dnf_installed_but_no_config_value_testcase = {
-            "legacy_type": 'RevertToImageDefault',
-            "stdio": {
-                "capture_output": False,
-                "expected_output": None
-            },
-            "config": {
-                "current_auto_update_config": {
-                    "create_current_auto_os_config": True,
-                    "current_auto_os_update_config_value": 'test_value = yes\n'
-                },
-                "backup_system_default_config": {
-                    "create_backup_for_system_default_config": True,
-                    "apply_updates_value": "",
-                    "download_updates_value": "",
-                    "enable_on_reboot_value": False,
-                    "installation_state_value": False,
-                    "set_installation_state": True
-                }
-            },
-            "assertions": {
-                "config_value_expected": 'download_updates =\napply_updates = \n',
-                "config_exists": True
-            }
-        }
-
-        revert_success_backup_config_does_not_exist_testcase = {
-            "legacy_type": 'RevertToImageDefault',
-            "stdio": {
-                "capture_output": True,
-                "expected_output": "[AzL3TDNF] Since the backup is invalid or does not exist for current service, we won't be able to revert auto OS patch settings to their system default value. [Service=dnf-automatic]"
-            },
-            "config": {
-                "current_auto_update_config": {
-                    "create_current_auto_os_config": True,
-                    "current_auto_os_update_config_value": 'apply_updates = no\ndownload_updates = no\n'
-                },
-                "backup_system_default_config": {
-                    "create_backup_for_system_default_config": False,
-                    "apply_updates_value": "",
-                    "download_updates_value": "",
-                    "enable_on_reboot_value": False,
-                    "installation_state_value": False,
-                    "set_installation_state": True
-                }
-            },
-            "assertions": {
-                "config_value_expected": 'apply_updates = no\ndownload_updates = no\n',
-                "config_exists": True
-            }
-        }
-
-        revert_success_default_backup_config_invalid_testcase = {
-            "legacy_type": 'RevertToImageDefault',
-            "stdio": {
-                "capture_output": True,
-                "expected_output": "[AzL3TDNF] Since the backup is invalid or does not exist for current service, we won't be able to revert auto OS patch settings to their system default value. [Service=dnf-automatic]"
-            },
-            "config": {
-                "current_auto_update_config": {
-                    "create_current_auto_os_config": True,
-                    "current_auto_os_update_config_value": 'apply_updates = no\ndownload_updates = no\n'
-                },
-                "backup_system_default_config": {
-                    "create_backup_for_system_default_config": True,
-                    "apply_updates_value": "yes",
-                    "download_updates_value": "yes",
-                    "enable_on_reboot_value": True,
-                    "installation_state_value": False,
-                    "set_installation_state": False
-                }
-            },
-            "assertions": {
-                "config_value_expected": 'apply_updates = no\ndownload_updates = no\n',
-                "config_exists": True
-            }
-        }
-
-        all_testcases = [revert_success_testcase, revert_success_with_dnf_not_installed_testcase, revert_success_with_dnf_installed_but_no_config_value_testcase, revert_success_backup_config_does_not_exist_testcase, revert_success_default_backup_config_invalid_testcase]
-
-        for testcase in all_testcases:
-            self.tearDown()
-            self.setUp()
-            captured_output, original_stdout = None, None
-            if testcase["stdio"]["capture_output"]:
-                # arrange capture std IO
-                captured_output, original_stdout = self.__capture_std_io()
-
-            self.runtime.set_legacy_test_type(testcase["legacy_type"])
-            package_manager = self.container.get('package_manager')
-
-            # setup current auto OS update config, backup for system default config and invoke revert to system default
-            self.__setup_config_and_invoke_revert_auto_os_to_system_default(package_manager,
-                                                                            create_current_auto_os_config=bool(testcase["config"]["current_auto_update_config"]["create_current_auto_os_config"]),
-                                                                            current_auto_os_update_config_value=testcase["config"]["current_auto_update_config"]["current_auto_os_update_config_value"],
-                                                                            create_backup_for_system_default_config=bool(testcase["config"]["backup_system_default_config"]["create_backup_for_system_default_config"]),
-                                                                            apply_updates_value=testcase["config"]["backup_system_default_config"]["apply_updates_value"],
-                                                                            download_updates_value=testcase["config"]["backup_system_default_config"]["download_updates_value"],
-                                                                            enable_on_reboot_value=bool(testcase["config"]["backup_system_default_config"]["enable_on_reboot_value"]),
-                                                                            installation_state_value=bool(testcase["config"]["backup_system_default_config"]["installation_state_value"]),
-                                                                            set_installation_state=bool(testcase["config"]["backup_system_default_config"]["set_installation_state"]))
-
-            # assert
-            if testcase["stdio"]["capture_output"]:
-                # restore sys.stdout output
-                sys.stdout = original_stdout
-                self.__assert_std_io(captured_output=captured_output, expected_output=testcase["stdio"]["expected_output"])
-            self.__assert_reverted_automatic_patch_configuration_settings(package_manager, config_exists=bool(testcase["assertions"]["config_exists"]), config_value_expected=testcase["assertions"]["config_value_expected"])
 
     def test_set_max_patch_publish_date(self):
         """Unit test for tdnf package manager set_max_patch_publish_date method"""
